Fix metric calculation
