+++ src/streamsight/metrics/binary/__init__.py
@@ -1,0 +1,4 @@
+from .hit import HitK
+
+
+__all__ = ["HitK"]
